FIX: Build under linux-arm
(cherry picked from commit e763ccc9165cf6cd140e6bf92d9cc2408caf6232)

+++ components/doublecmd/dcunix.pas
@@ -213,7 +213,7 @@ begin
 {$IFDEF LINUX}
   if (fpFStatFS(Handle, @Sbfs) = 0) then
   begin
-    case Sbfs.fstype of
+    case UInt32(Sbfs.fstype) of
       SMB_SUPER_MAGIC,
       SMB2_MAGIC_NUMBER,
       CIFS_MAGIC_NUMBER: Exit;
